% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{build_graph_from_files}
\alias{build_graph_from_files}
\title{Build a function-dependencies graph for a list of files}
\usage{
build_graph_from_files(file_ls)
}
\arguments{
\item{file_ls}{Character vector of file paths.}
}
\description{
Build a function-dependencies graph for a list of files
}
