% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{build_pkg_graph}
\alias{build_pkg_graph}
\title{Build a graph for a package}
\usage{
build_pkg_graph()
}
\description{
A one-line call to build a graph from a package.
}
\details{
This function binds 'build_graph_from_dir', 'remove_duplicates',
'remove_vbar', 'prepare_graph_Rgraphviz' and 'start_app' into one call.
}
