% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cidr.R
\name{compute_cidr}
\alias{compute_cidr}
\title{Cumulative Intraday Returns}
\usage{
compute_cidr(dat)
}
\arguments{
\item{dat}{Functional data as a matrix with the columns as individual
observations.}
}
\value{
Functional data for the CIDRs
}
\description{
Compute the cumulative intraday returns.
}
\examples{
compute_cidr(matrix(runif(5000),ncol=100))
}
\references{
ice, G., Wirjanto, T., & Zhao, Y. (2019). Tests for conditional
heteroscedasticity with functional data and goodness-of-fit tests for FGARCH
models. IDEAS Working Paper Series from RePEc.
}
