% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_change.R
\name{cov_change}
\alias{cov_change}
\title{Covariance Change}
\usage{
cov_change(X, kappa = 1/4, len = 30)
}
\arguments{
\item{X}{Numeric data.frame of functional data observations--rows for
evaluated values and columns indicating FD}

\item{kappa}{(Optional) Numeric used for weighting and such. Default is 1/4.}

\item{len}{(Optional) Numeric for window/repetitions for covariance change.
Default is 30.}
}
\value{
Location of change. NA for no change and numeric if there is a change.
}
\description{
This method implements a method for detection of covariance changes in
functional data.
}
\references{
Change point analysis of covariance functions: a weighted
cumulative sum approach, L. Horvath, G. Rice, Y. Zhao (2022) Journal of
Multivariate Analysis.
}
