% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{electricity}
\alias{electricity}
\title{Spanish Spot Electricity Data}
\format{
\subsection{\code{electricity}}{

A data.frame with columns being the 365 days and rows the 24 hours.
}
}
\source{
<www.omie.es>
}
\usage{
electricity
}
\description{
The hourly electricity spot prices from Spain in 2014.
}
\keyword{datasets}
