% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_change.R
\name{graph_segmentation}
\alias{graph_segmentation}
\title{Graph-based change point detection}
\usage{
graph_segmentation(
  n,
  E,
  statistics = c("all", "o", "w", "g", "m"),
  n0 = 0.05 * n,
  n1 = 0.95 * n,
  pval.appr = TRUE,
  skew.corr = TRUE,
  pval.perm = FALSE,
  B = 100,
  silent = FALSE
)
}
\arguments{
\item{n}{length of sample}

\item{E}{edge data from a graph}

\item{statistics}{Statistic to select method}

\item{n0, n1}{Starting/ending value (for trimming)}

\item{pval.appr}{Boolean if pvalue approximation should be used}

\item{skew.corr}{Boolean if skewness correction should be used with pvalue
approximation}

\item{pval.perm}{Boolean if pvalue permutation should be used}

\item{B}{Number of simulations for null distribution}

\item{silent}{Boolean if output should be displayed}
}
\value{
List of graph change details
}
\description{
Graph-based change point detection
}
\references{
Chen H, Zhang NR, Chu L, Song H (2020). \emph{gSeg: Graph-Based
Change-Point Detection (g-Segmentation)}. R package version 1.0.
}
