% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{pedestrians}
\alias{pedestrians}
\title{Australian Pedestrians}
\format{
\subsection{\code{pedestrians}}{

A data.frame with columns being the 312 days and rows the 24 hours.
}
}
\usage{
pedestrians
}
\description{
Number of pedestrians at an intersection in Melbourne Australia (2019-2021).
}
\keyword{datasets}
