% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{prepare_graph}
\alias{prepare_graph}
\title{Add plotting parameters to the graph data}
\usage{
prepare_graph(l0, layout)
}
\arguments{
\item{l0}{A pairlist of edgelist and nodelist.}

\item{layout}{(Optional) A 2-column matrix of nodes coordinates.}
}
\description{
Add plotting parameters to the graph data
}
