% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{prepare_graph_Rgraphviz}
\alias{prepare_graph_Rgraphviz}
\title{Add plotting parameters to the graph data}
\usage{
prepare_graph_Rgraphviz(l0, ...)
}
\arguments{
\item{l0}{A pairlist of edgelist and nodelist.}

\item{...}{Extra scaling parameters to pass to `snap_to_grid`.}
}
\description{
Add plotting parameters to the graph data
}
\examples{
\dontrun{
library(funGraphs)
library(magrittr)
build_graph_from_dir("R/") \%>\%
  prepare_graph_Rgraphviz() \%>\%
  start_app()
}
}
