% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_vbar}
\alias{remove_vbar}
\title{[Hot-fix] Remove entries with "|" symbol}
\usage{
remove_vbar(l0)
}
\arguments{
\item{l0}{A pairlist of edgelist and nodelist.}
}
\description{
igraphs does not allow the symbol "|" in a node's name.
Entries with such symbol are removed.
}
\examples{
\dontrun{
build_graph_from_dir("R/") \%>\%
  remove_vbar() \%>\%
  prepare_graph_Rgraphviz() \%>\%     # Add plotting parameters
  start_app()
}
}
