% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{start_app}
\alias{start_app}
\title{Plotting the package dependencies graph}
\usage{
start_app(l0, browser = getOption("viewer"))
}
\arguments{
\item{l0}{A pairlist of edgelist and nodelist.}

\item{browser}{Use "browser" or "viewer" to view the graph.}
}
\description{
Plotting the package dependencies graph
}
