% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{TWTR}
\alias{TWTR}
\title{Twitter Stock Prices}
\format{
\subsection{\code{rates}}{

A data.frame with columns being the 505 days and rows the 15 minute intraday time.
}
}
\usage{
TWTR
}
\description{
Stock prices for Twitter from 2019 to 2020)
(removing days without information, i.e. weekends and holidays).
}
\keyword{datasets}
