% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{update_graph_from_svg}
\alias{update_graph_from_svg}
\title{Update graph coordinates from a SVG file}
\usage{
update_graph_from_svg(l0, svg_file)
}
\arguments{
\item{l0}{A pairlist of edgelist and nodelist; the graph.}

\item{svg_file}{character; path to the svg file.}
}
\description{
Update graph coordinates from a SVG file
}
