% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{build_graph_from_dir}
\alias{build_graph_from_dir}
\title{Build a function-dependencies graph for an R package}
\usage{
build_graph_from_dir(dir0)
}
\arguments{
\item{dir0}{A directory path.}
}
\description{
Build a function-dependencies graph for an R package
}
